WS-15: Organize modules into Feature modules and shared modules
