+++ client/src/app/features/main-dashboard/global-insights/global-insights.component.less
